тип ingredientBox
